Add files via upload
This notebook is inspired by the work of Elholm et al. (2022) DOI: 10.1039/d2cp03032b from the group of Kasper Moth-Poulsen working on molecular solar thermal energy storage. The background on clustering was based on Greg Landrum's work as well as the Liam Wilbraham for substituent generation. The libraries rdkit and ase are used. The code for python to sql conversions was taken from the tutorial by Craig Dickson and uses an MySQL Server.