Add files via upload
This notebook is inspired by the work of Elholm et al. (2022) DOI: 10.1039/d2cp03032b from the group of Kasper Moth-Poulsen working on molecular solar thermal energy storage. Here, the python packages rdkit and ase are used for chemical calculations.
The code for python to sql conversions is based on the tutorial by Craig Dickson and uses an MySQL Server.